# Remove allocations in BaseUriHelper, improve resource resolution performance