--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/PresentationCore.csproj
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/PresentationCore.csproj
@@ -344,6 +344,7 @@
     <Compile Include="System\Windows\BinaryFormat\ListConverterHelper.cs" />
     <Compile Include="System\Windows\BinaryFormat\FormatterConverterStub.cs" />
     <Compile Include="System\Windows\BinaryFormat\SerializationExtensions.cs" />
+    <Compile Include="System\Windows\Navigation\AssemblyPackageInfo.cs" />
     <Compile Include="System\Windows\Nrbf\SerializationRecordExtensions.cs" />
     <Compile Include="System\Windows\BinaryFormat\MessageEnd.cs" />
     <Compile Include="System\Windows\BinaryFormat\Record.cs" />