--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/Effects/PixelShader.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/Effects/PixelShader.cs
@@ -1,4 +1,4 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
@@ -76,7 +76,7 @@ namespace System.Windows.Media.Effects
                 {
                     if (!newUri.IsAbsoluteUri)
                     {
-                         newUri = BaseUriHelper.GetResolvedUri(BaseUriHelper.BaseUri, newUri);
+                         newUri = BaseUriHelper.GetResolvedUri(BaseUriHelper.PackAppBaseUri, newUri);
                     }
 
                     Debug.Assert(newUri.IsAbsoluteUri);