--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/Data/DefaultValueConverter.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/Data/DefaultValueConverter.cs
@@ -1,4 +1,4 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
@@ -759,7 +759,7 @@ namespace MS.Internal.Data
                     }
                     else
                     {
-                        _cachedBaseUri = BaseUriHelper.BaseUri;
+                        _cachedBaseUri = BaseUriHelper.PackAppBaseUri;
                     }
                 }
                 return _cachedBaseUri;