--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Application.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Application.cs
@@ -1952,10 +1952,7 @@ namespace System.Windows
         {
             get
             {
-                if (_applicationMarkupBaseUri == null)
-                {
-                    _applicationMarkupBaseUri = BaseUriHelper.BaseUri;
-                }
+                _applicationMarkupBaseUri ??= BaseUriHelper.PackAppBaseUri;
 
                 return _applicationMarkupBaseUri;
             }