--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Data/XmlDataProvider.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Data/XmlDataProvider.cs
@@ -23,6 +23,7 @@ using System.Windows.Markup; // IUriContext, [XamlDesignerSerializer]
 using MS.Internal;                  // CriticalExceptions
 using MS.Internal.Utility;          // BindUriHelper
 using MS.Internal.Data;             // XmlDataCollection
+using System.Windows.Navigation;
 
 namespace System.Windows.Data
 {
@@ -350,9 +351,9 @@ namespace System.Windows.Data
         {
             // convert the Source into an absolute URI
             Uri sourceUri = this.Source;
-            if (sourceUri.IsAbsoluteUri == false)
+            if (!sourceUri.IsAbsoluteUri)
             {
-                Uri baseUri = _baseUri ?? BindUriHelper.BaseUri;
+                Uri baseUri = _baseUri ?? BaseUriHelper.PackAppBaseUri;
                 sourceUri = BindUriHelper.GetResolvedUri(baseUri, sourceUri);
             }