--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Markup/XamlReader.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Markup/XamlReader.cs
@@ -1137,15 +1137,15 @@ namespace System.Windows.Markup
             return (root);
         }
 
-        static Uri GetBaseUri(Uri uri)
+        private static Uri GetBaseUri(Uri uri)
         {
-            if (uri == null)
+            if (uri is null)
             {
-                return MS.Internal.Utility.BindUriHelper.BaseUri;
+                return BaseUriHelper.PackAppBaseUri;
             }
-            else if (uri.IsAbsoluteUri == false)
+            else if (!uri.IsAbsoluteUri)
             {
-                return new Uri(MS.Internal.Utility.BindUriHelper.BaseUri, uri);
+                return new Uri(BaseUriHelper.PackAppBaseUri, uri);
             }
 
             return uri;