--- a/src/Microsoft.DotNet.Wpf/src/Shared/MS/Internal/ReflectionUtils.cs
+++ b/src/Microsoft.DotNet.Wpf/src/Shared/MS/Internal/ReflectionUtils.cs
@@ -5,10 +5,9 @@
 
 using System.Runtime.CompilerServices;
 using System.Reflection.Metadata;
-using System.Reflection;
-using System.Text;
-using System;
 using System.Diagnostics;
+using System.Reflection;
+using System;
 
 namespace MS.Internal
 {
@@ -124,6 +123,20 @@ namespace MS.Internal
                 if (assemblyToken.Length != 16)
                     assemblyToken = ReadOnlySpan<char>.Empty;
             }
+        }
+
+        /// <summary>
+        /// Compares whether two PublicKeyTokens are the same.
+        /// If both <paramref name="reqToken"/> and <paramref name="curToken"/> are <see langword="null"/>, <see langword="true"/> is returned.
+        /// </summary>
+        /// <param name="reqToken">First PublicKeyToken to compare.</param>
+        /// <param name="curToken">Second PublicKeyToken to compare.</param>
+        /// <returns><see langword="true"/> if both parameters are <see langword="null"/> or equal in sequence, <see langword="false"/> otherwise.</returns>
+        internal static bool IsSamePublicKeyToken(byte[] reqToken, byte[] curToken)
+        {
+            Debug.Assert((reqToken is null || reqToken.Length is 0 or 8) && (curToken is null || curToken.Length is 0 or 8), "Provided PublicKeyToken has invalid length");
+
+            return (reqToken is null && curToken is null) || (curToken is not null && reqToken is not null && reqToken.AsSpan().SequenceEqual(curToken));
         }
 
         /// <summary>