Currently blocked by #9822, hence draft.

## Description

Optimizes methods in `BaseUriHelper`, removes most of the allocations directly in this class and opens up more optimization possibilities; directly builds upon #9739 and #9822.

- Removes `IsSameKeyToken` from SecurityHelper in favour of `ReflectionUtils.IsSamePublicKeyToken`.
- `BaseUriHelper.GetAssemblyNameAndPart` is now fully allocation free.
    - This allows for more optimizations in the future and already makes some paths without any allocations.
    - Instead of returning 3 strings, we return slices in `AssemblyPackageInfo` `ref struct`.
- In `FontResourceCache`, we can now use `AlternateLookup` to benefit from this change as well.
- `BaseUriHelper.BaseUri` was never set outside static constructor, where it was initialized to the `s_packAppBaseUri` value,
hence I've removed the property altogether, which also led to removal of `BindUriHelper.BaseUri` which was just proxy.
- Code under `CF_Envelope_Activation_Enabled` macro was removed, most of the code is already missing in .NET Core than what was in reference files for .NET 3.5 and it is never compiled on NetFX nor on .NET Core.
- Changes in `GetLoadedAssembly` bring a bit of an improvement but there's more benchmarking needed on how to handle this chain most efficiently, so keeping as it is right now.

Cherry-picking some cases:

### GetAssemblyNameAndPart

| Method   | someUri             | Mean [ns] | Error [ns] | StdDev [ns] |  Gen0   | Allocated [B] |
|--------- |-------------------- |----------:|-----------:|------------:|---------:|--------------:|
| Original | /Mic(...)xaml [103] | 104.95 ns |   2.274 ns |    6.706 ns |     0.0448 |         752 B |
| PR__EDIT | /Mic(...)xaml [103] |  45.42 ns |   0.475 ns |    0.445 ns |         - |             - |

### AssemblyMatchesKeyString vs ComparePublicKeyTokens

| Method   | someAssembly         | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Allocated [B] |
|--------- |--------------------- |----------:|-----------:|------------:|-------:|---------------:|
| Original | Syste(...)7798e [89] | 40.755 ns |  0.3990 ns |   0.3537 ns | 0.0019 |          32 B |
| PR__EDIT | Syste(...)7798e [89] |  8.135 ns |  0.0835 ns |   0.0740 ns |      - |                 - |

## Customer Impact

Improved performance, decrased allocations.

## Regression

No.

## Testing

Local build, some assert testing; I'll provide tests before this goes out of draft.

## Risk

Low to medium, there are quite a few changes but most of them are rather straightforward.